moved customer routes to customerController
