+++ src/main/java/com/company/RareBookStore/controller/BookController.java
@@ -1,9 +1,7 @@
 package com.company.RareBookStore.controller;
 
 import com.company.RareBookStore.model.Book;
-import com.company.RareBookStore.model.Customer;
 import com.company.RareBookStore.repository.BookRepository;
-import com.company.RareBookStore.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +14,6 @@ public class BookController {
 
     @Autowired
     private BookRepository bookRepository;
-    private CustomerRepository customerRepository;
 
     //GET ALL BOOKS
     @GetMapping("/books")
@@ -26,27 +23,11 @@ public class BookController {
         return bookRepository.findAll();
     }
 
-    //GET ALL CUSTOMERS
-    @GetMapping("/customers")
-    @ResponseStatus(HttpStatus.OK)
-    public List<Customer> getAllCustomers() {
-
-        return customerRepository.findAll();
-    }
-
     //GET BOOK BY ID
     @GetMapping("/books/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Book getBookById(@PathVariable Integer id) {
         Optional<Book> returnVal = bookRepository.findById(id);
-        return returnVal.orElse(null);
-    }
-
-    //GET CUSTOMER BY ID
-    @GetMapping("/customers/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public Customer getCustomerById(@PathVariable Integer id) {
-        Optional<Customer> returnVal = customerRepository.findById(id);
         return returnVal.orElse(null);
     }
 
@@ -89,16 +70,6 @@ public class BookController {
         bookRepository.save(book);
     }
 
-    //UPDATE CUSTOMER BY ID
-    @PutMapping("/customers/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void updateCustomer(@RequestBody Customer customer, @PathVariable Integer id) {
-        if(customer.getId() != id) {
-            throw new IllegalArgumentException("Entered ID does not match existing customer ID");
-        }
-        customerRepository.save(customer);
-    }
-
     //DELETE BOOK BY ID
     @DeleteMapping("/books/{id}")
     @ResponseStatus(HttpStatus.OK)
@@ -106,12 +77,6 @@ public class BookController {
         bookRepository.deleteById(id);
     }
 
-    //DELETE CUSTOMER BY ID
-    @DeleteMapping("/customers/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteCustomer(@RequestBody Customer customer, @PathVariable Integer id) {
-        customerRepository.deleteById(id);
-    }
 }
